fixed navbar colors, added eye icon on login to allow users to see password
