+++ remi-app/app/poo.tsx
@@ -1,12 +1,11 @@
-import { useRouter } from "expo-router"; // Add this import
-import { useEffect, useState } from "react";
+import { useRouter } from "expo-router";
+import { useEffect, useState, useCallback } from "react";
 import {
   Text,
   View,
   StyleSheet,
   KeyboardAvoidingView,
   TextInput,
-  Button,
   ActivityIndicator,
   ImageBackground,
   TouchableWithoutFeedback,
@@ -20,7 +19,6 @@ import {
 import { FirebaseError } from "firebase/app";
 import { auth } from "../firebaseConfig";
 import * as Font from "expo-font";
-import AppLoading from "expo-app-loading";
 import {
   useFonts,
   OrelegaOne_400Regular,
@@ -31,13 +29,18 @@ import {
   Nunito_700Bold,
 } from "@expo-google-fonts/nunito";
 import Ustyles from "../components/UniversalStyles";
-// export var isCreateAccount = useState(false);
+import * as SplashScreen from "expo-splash-screen";
+import { Ionicons } from "@expo/vector-icons";
+
+SplashScreen.preventAutoHideAsync();
 
 export default function Index() {
-  const router = useRouter(); // Initialize router
+  const router = useRouter();
   const [email, setEmail] = useState("");
   const [password, setPassword] = useState("");
   const [loading, setLoading] = useState(false);
+  const [appIsReady, setAppIsReady] = useState(false);
+  const [showPassword, setShowPassword] = useState(false);
 
   const signIn = async () => {
     setLoading(true);
@@ -51,6 +54,10 @@ export default function Index() {
     }
   };
 
+  const togglePasswordVisibility = () => {
+    setShowPassword(!showPassword);
+  };
+
   let [fontsLoaded] = useFonts({
     OrelegaOne_400Regular,
     Nunito_700Bold,
@@ -58,8 +65,20 @@ export default function Index() {
     Nunito_400Regular,
   });
 
-  if (!fontsLoaded) {
-    return <AppLoading />;
+  // Prepare the app and hide the splash screen once fonts are loaded
+  useEffect(() => {
+    async function prepare() {
+      if (fontsLoaded) {
+        await SplashScreen.hideAsync();
+        setAppIsReady(true);
+      }
+    }
+    prepare();
+  }, [fontsLoaded]);
+
+  // Avoid rendering the main content until the splash screen is hidden
+  if (!appIsReady) {
+    return null;
   }
 
   return (
@@ -88,15 +107,27 @@ export default function Index() {
                 placeholderTextColor="#BCD5AC"
                 placeholder="Email"
               />
-              <TextInput
-                style={styles.input}
-                value={password}
-                onChangeText={setPassword}
-                autoCorrect={false}
-                secureTextEntry
-                placeholder="Password"
-                placeholderTextColor="#BCD5AC"
-              />
+              <View style={styles.passwordContainer}>
+                <TextInput
+                  style={styles.passwordInput}
+                  value={password}
+                  onChangeText={setPassword}
+                  autoCorrect={false}
+                  secureTextEntry={!showPassword}
+                  placeholder="Password"
+                  placeholderTextColor="#BCD5AC"
+                />
+                <TouchableOpacity
+                  onPress={togglePasswordVisibility}
+                  style={styles.eyeIcon}
+                >
+                  <Ionicons
+                    name={showPassword ? "eye-off-outline" : "eye-outline"}
+                    size={24}
+                    color="#0D5F13"
+                  />
+                </TouchableOpacity>
+              </View>
 
               {loading ? (
                 <ActivityIndicator size={"small"} style={{ margin: 28 }} />
@@ -105,14 +136,13 @@ export default function Index() {
                   <TouchableOpacity style={Ustyles.button} onPress={signIn}>
                     <Text style={Ustyles.header_2}>Sign In</Text>
                   </TouchableOpacity>
-                  {/* Navigate to Register Page */}
                   <TouchableOpacity
                     style={Ustyles.button}
                     onPress={() => router.push("./(auth)/register")}
                   >
                     <Text style={Ustyles.header_2}>Create Account</Text>
                   </TouchableOpacity>
-                </View> // this was not working for rainas laptop idk
+                </View>
               )}
             </KeyboardAvoidingView>
           </View>
@@ -121,6 +151,7 @@ export default function Index() {
     </TouchableWithoutFeedback>
   );
 }
+
 const styles = StyleSheet.create({
   container: {
     marginHorizontal: 20,
@@ -135,5 +166,22 @@ const styles = StyleSheet.create({
     padding: 10,
     backgroundColor: "#fff",
     borderColor: "#0D5F13",
+  },
+  passwordContainer: {
+    flexDirection: "row",
+    alignItems: "center",
+    marginVertical: 4,
+    borderWidth: 2,
+    borderRadius: 4,
+    backgroundColor: "#fff",
+    borderColor: "#0D5F13",
+  },
+  passwordInput: {
+    flex: 1,
+    height: 50,
+    padding: 10,
+  },
+  eyeIcon: {
+    padding: 10,
   },
 });
